requirements.txt
# Requirements for Lakebase Training Dashboard
# Python 3.9+

# Web Framework
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# Database
psycopg[binary]>=3.1.0

# Data processing
pandas>=2.0.0

# Databricks SDK
databricks-sdk>=0.18.0
